Bug Fix: Correct celerybeat init.d defaults
The CELERYBEAT variable in the celerybeat default file is required to be
wrapped in " to work. This was being ignored when not in ". This commit
wraps the parameter in ".

The init.d for celerybeat was referencing the default file with a .conf
extension, but the default file was not actually created with a .conf
extension. This commit removes the .conf from the init.d file.